Initial move to specs dir
