+++ datadog_checks_dev/datadog_checks/dev/tooling/commands/validate/config.py
@@ -6,8 +6,8 @@ import yaml
 
 from datadog_checks.dev.tooling.config_validator.validator import validate_config
 from datadog_checks.dev.tooling.config_validator.validator_errors import SEVERITY_ERROR, SEVERITY_WARNING
-from datadog_checks.dev.tooling.configuration import ConfigSpec
-from datadog_checks.dev.tooling.configuration.consumers import ExampleConsumer
+from datadog_checks.dev.tooling.specs.configuration import ConfigSpec
+from datadog_checks.dev.tooling.specs.configuration.consumers import ExampleConsumer
 
 from ....utils import basepath, file_exists, path_join, read_file, write_file
 from ...utils import (
