+++ datadog_checks_dev/datadog_checks/dev/tooling/specs/configuration/consumers/example.py
