+++ datadog_checks_dev/datadog_checks/dev/tooling/specs/configuration/template.py
@@ -5,9 +5,9 @@ from copy import deepcopy
 
 import yaml
 
-from ...utils import file_exists, get_parent_dir, path_join, read_file
+from ....utils import file_exists, get_parent_dir, path_join, read_file
 
-TEMPLATES_DIR = path_join(get_parent_dir(get_parent_dir(__file__)), 'templates', 'configuration')
+TEMPLATES_DIR = path_join(get_parent_dir(get_parent_dir(get_parent_dir(__file__))), 'templates', 'configuration')
 VALID_EXTENSIONS = ('yaml', 'yml')
