+++ datadog_checks_dev/tests/tooling/configuration/test_template.py
@@ -6,7 +6,7 @@ import re
 import pytest
 
 from datadog_checks.dev import TempDir
-from datadog_checks.dev.tooling.configuration.template import ConfigTemplates
+from datadog_checks.dev.tooling.specs.configuration.template import ConfigTemplates
 from datadog_checks.dev.utils import ensure_parent_dir_exists, path_join, write_file
 
 pytestmark = [pytest.mark.conf, pytest.mark.conf_template]
