+++ datadog_checks_dev/tests/tooling/configuration/utils.py
@@ -3,8 +3,8 @@
 # Licensed under a 3-clause BSD style license (see LICENSE)
 from textwrap import dedent
 
-from datadog_checks.dev.tooling.configuration import ConfigSpec
-from datadog_checks.dev.tooling.configuration.consumers import ExampleConsumer
+from datadog_checks.dev.tooling.specs.configuration import ConfigSpec
+from datadog_checks.dev.tooling.specs.configuration.consumers import ExampleConsumer
 
 
 def get_spec(text, **kwargs):
